requirements.txt
transformers
torch
streamlit
python-dotenv
pandas
plotly
